Resolve -Wsign-conversion Warning Messages:
Building with Project Flag -DALL_WARNINGS=True, compiler flagged
numerous lines where there were implicit value conversions. These lines
were function calls that had arguments, for the most part, expecting
integer sign change.

Passed values were explicitly converted (static_cast). Where possible,
tests of variables were added to ensure contents were acceptable to
function call.

If problem occurred with QVariant.toInt() call, the .toUInt() function
was used instead. Note: this does not resolve whether QVariant stores
a negative value in the first place.

Std keyword `auto` replaced when compiler is unable to determine type
to use because of multiple types in assignment statement.

NOTE: -Wsign-conversion warnings in the following files were untouched:
[project root]/src/RowCache.h
[project root]/src/csvparser.cpp
[project root]/src/sqlitedb.cpp
[project root]/src/grammar/Sqlite3Lexer.cpp
(previously discussed in #1788)
[project root]/src/grammar/Sqlite3Parser.cpp
(previously discussed in #1788)

Note: index accessor to std::vector elements uses size_type
<aka long unsigned int>
while QVector element accessors utilize integer.

+++ src/ForeignKeyEditorDelegate.cpp
@@ -123,7 +123,7 @@ void ForeignKeyEditorDelegate::setEditorData(QWidget* editor, const QModelIndex&
     ForeignKeyEditor* fkEditor = static_cast<ForeignKeyEditor*>(editor);
 
     int column = index.row(); // weird? I know right
-    const sqlb::Field& field = m_table.fields.at(column);
+    const sqlb::Field& field = m_table.fields.at(static_cast<size_t>(column));
     auto fk = std::dynamic_pointer_cast<sqlb::ForeignKeyClause>(m_table.constraint({field.name()}, sqlb::Constraint::ForeignKeyConstraintType));
     if (fk) {
         fkEditor->tablesComboBox->setCurrentText(fk->table());
@@ -141,7 +141,7 @@ void ForeignKeyEditorDelegate::setModelData(QWidget* editor, QAbstractItemModel*
     QString sql = fkEditor->getSql();
 
     int column = index.row();
-    const sqlb::Field& field = m_table.fields.at(column);
+    const sqlb::Field& field = m_table.fields.at(static_cast<size_t>(column));
     if (sql.isEmpty()) {
         // Remove the foreign key
         m_table.removeConstraints({field.name()}, sqlb::Constraint::ConstraintTypes::ForeignKeyConstraintType);
